+++ pkg/cmd/inventoryFlags.go
@@ -109,17 +109,9 @@ func getTenantWithDefaultFlag(cmd *cobra.Command, flagName string, defaultTenant
 	return defaultTenant
 }
 
-func getFormDataObjectFlag(cmd *cobra.Command, flagName string, formData map[string]io.Reader) error {
-	if formData == nil {
-		return fmt.Errorf("formData can not be nil")
-	}
-
-	if value, err := cmd.Flags().GetString(FlagDataName); err == nil {
-		data := MustParseJSON(value)
-
-		if metadataBytes, err := json.Marshal(data); err == nil {
-			formData["object"] = bytes.NewReader(metadataBytes)
-		}
+func getFormDataObjectFlag(cmd *cobra.Command, flagName string, data map[string]interface{}) error {
+	if value, err := cmd.Flags().GetString(flagName); err == nil {
+		return ParseJSON(value, data)
 	}
 	return nil
 }
@@ -130,27 +122,40 @@ func getFileFlag(cmd *cobra.Command, flagName string, formData map[string]io.Rea
 	}
 
 	// Get custom properties which should be added to the binary
-	getFormDataObjectFlag(cmd, FlagDataName, formData)
+	objectInfo := make(map[string]interface{})
+	err := getFormDataObjectFlag(cmd, FlagDataName, objectInfo)
+	if err != nil {
+		return newSystemErrorF("Could not parse %s flag. %s", FlagDataName, err)
+	}
 
 	if filename, err := cmd.Flags().GetString(flagName); err == nil {
 		r, err := os.Open(filename)
-		if err == nil {
-			formData["file"] = r
-
-			// Add required object field if it does not already exist
-			if _, ok := formData["object"]; !ok {
-				objectInfo := make(map[string]interface{})
-				objectInfo["type"] = mime.TypeByExtension(filepath.Ext(filename))
-				objectInfo["name"] = filepath.Base(filename)
-				if v, err := json.Marshal(objectInfo); err == nil {
-					formData["object"] = bytes.NewReader(v)
-				} else {
-					return errors.New("failed to create object form-data property")
-				}
-			}
-		} else {
+		if err != nil {
 			return errors.New("Failed to read file")
 		}
+
+		formData["file"] = r
+
+		if _, ok := objectInfo["name"]; !ok {
+			objectInfo["name"] = filepath.Base(filename)
+		}
+
+		if _, ok := objectInfo["type"]; !ok {
+			mimeType := mime.TypeByExtension(filepath.Ext(filename))
+
+			if mimeType == "" {
+				mimeType = "application/octet-stream"
+			}
+
+			objectInfo["type"] = mimeType
+		}
+
+		if v, err := json.Marshal(objectInfo); err == nil {
+			formData["object"] = bytes.NewReader(v)
+		} else {
+			return errors.New("failed to create object form-data property")
+		}
+
 	}
 	return nil
 }